adds the stuff
