+++ x-pack/plugins/endpoint/common/alerts/schema/alert_index.ts
