+++ x-pack/plugins/endpoint/common/alerts/schema/index_pattern.ts
