--- x-pack/plugins/endpoint/common/schema/README.md
@@ -1,6 +1,0 @@
-# Schemas
-
-These schemas are used to validate, coerce, and provide types for the comms between the client, server, and ES.
-
-# Future work
-In the future, we may be able to locate these under 'server'.
