+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/models/index_pattern.ts
