+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/store/action.ts
