+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/store/middleware.ts
