+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/store/mock_alert_result_list.ts
