+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/details/index.ts
