+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/details/metadata/file_accordion.tsx
