+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/details/metadata/general_accordion.tsx
