+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/details/metadata/source_process_token_accordion.tsx
