+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/details/overview/take_action_dropdown.tsx
