+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/formatted_date.tsx
