+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/resolver.tsx
