+++ x-pack/plugins/endpoint/public/applications/endpoint/alerts/view/url_from_query_params.ts
